requirements.txt
feedparser
requests
beautifulsoup4
PyYAML
python-dotenv
google-generativeai
# 필요한 경우 여기에 생성형 AI 라이브러리 추가 (예: google-generativeai, openai) 